Moving towards making msgpack optional.
