+++ Tensile/Source/lib/CMakeLists.txt
@@ -58,6 +58,12 @@ if(TENSILE_USE_LLVM)
     )
 endif()
 
+if(TENSILE_USE_MSGPACK)
+    set(tensile_sources ${tensile_sources}
+        source/msgpack/MessagePack.cpp
+    )
+endif()
+
 if(TENSILE_USE_HIP)
     set(tensile_sources ${tensile_sources}
         source/hip/HipSolutionAdapter.cpp
@@ -71,11 +77,12 @@ add_library (TensileHost STATIC ${tensile_sources})
 
 target_include_directories(TensileHost INTERFACE "${CMAKE_CURRENT_SOURCE_DIR}/include")
 
-find_package(msgpack)
-target_link_libraries(TensileHost PUBLIC msgpackc)
+if(TENSILE_USE_LLVM OR TENSILE_USE_MSGPACK)
+    target_compile_definitions(TensileHost PUBLIC TENSILE_DEFAULT_SERIALIZATION)
+endif()
 
 if(TENSILE_USE_LLVM)
-    target_compile_definitions(TensileHost PUBLIC TENSILE_DEFAULT_SERIALIZATION)
+    target_compile_definitions(TensileHost PUBLIC -DTENSILE_YAML=1)
     find_library(LLVMObjectYAML_LIBRARY
         NAMES LLVMObjectYAML
         PATHS ${LLVM_LIBRARY_DIR})
@@ -89,6 +96,13 @@ if(TENSILE_USE_LLVM)
     target_include_directories(TensileHost PRIVATE ${LLVM_INCLUDE_DIRS})
 endif()
 
+if(TENSILE_USE_MSGPACK)
+    target_compile_definitions(TensileHost PUBLIC -DTENSILE_MSGPACK=1)
+
+    find_package(msgpack)
+    target_link_libraries(TensileHost PUBLIC msgpackc)
+endif()
+
 if(TENSILE_STATIC_ONLY)
     target_compile_definitions(TensileHost PUBLIC TENSILE_STATIC_ONLY)
 endif()
@@ -99,10 +113,6 @@ if(TENSILE_USE_HIP)
     #add_subdirectory(client)
     target_link_libraries( TensileHost PUBLIC ${HSA_LIBRARIES} hip::device )
 
-endif()
-
-if(Tensile_LIBRARY_FORMAT MATCHES "yaml")
-    target_compile_definitions(TensileHost PUBLIC -DTENSILE_YAML=1)
 endif()
 
 if(false)
